% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inR.activity.r
\name{incR.activity}
\alias{incR.activity}
\title{Calculation of onset and end of daily activity.}
\usage{
incR.activity(data, vector.presence)
}
\arguments{
\item{data:}{data frame containing a time-series vector of 1s and 0s, where "1"
means "incubating individual inside nest" and "0" means "incubating individual 
outside the nests". This vector, 
under the name of "inc.vector" is provided by \code{\link{incR.scan}} in the 
first object of the returned list. A column named "date" is needed to refer to daily
calculations.}

\item{vector.presence:}{(character class) name of the vector containing information
for incubating individual presence/absence of nests. \code{\link{incR.scan}} produces
this vector named "inc.vector".}
}
\value{
a data frame containing a column (vector) for incubating individual presence/
ausence of nests.
}
\description{
Using a vector of incubating individual presence in nest, 
\emph{incR.activity}
calculates onset of daily activity (first off-bout of a day),
and end of daily activity (last on-bout of a day). A column for dates, named "date"
is needed in the data argument.
}
\examples{
To be included
}
\author{
Pablo Capilla
}
\seealso{
\code{\link{incR.prep}} \code{\link{incR.scan}}
}

