% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inR.bouts.r
\name{incR.bouts}
\alias{incR.bouts}
\title{Calculation of the number of daily on- and off-bouts}
\usage{
incR.bouts(data, vector, sampling.rate)
}
\arguments{
\item{data}{ny}

\item{vector}{ny}

\item{sampling.rate}{ny}
}
\value{
ny
}
\description{
Calculation of
}
\examples{
To be included
}
\author{
Pablo Capilla
}
\seealso{
\code{\link{incR.prep}} \code{\link{incR.scan}} \code{\link{incR.activity}}
\code{\link{incR.constancy}}
}

