% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR.incubationExample.R
\docType{data}
\name{incR.incubationExample}
\alias{incR.incubationExample}
\title{An example of incubation temperature time-series after 
the use of \code{link{incR.scan}}.}
\format{A data frame with 954 rows and 10 variables, representing three days of nest
temperatures.
\describe{
  \item{DATE}{a date + time column. Both elements are concatenated in one column.}
  \item{valueT}{temperature recordings.}
  \item{index}{running number from first to last observation.}
  \item{time}{}
  \item{hour}{}
  \item{year}{}
  \item{date}{}
  \item{dec.time}{time in decimal hours.}
  \item{temp1}{difference between the \emph{ith} and the \emph{ith-1} 
  temperature recording.}
  \item{inc.vector}{ a vector-class object containing '1s' or '0s' depending on whether
  the incubating individual was calculated to be inside or ourside the nest.}
}}
\usage{
incR.incubationExample
}
\description{
A dataset containing temperatures, time and date for a blue tit nest. 
It also includes variables added by \code{\link{incR.prep}} and the
\emph{inc.vector} calculated by \code{\link{incR.scan}}. 
The raw data set \code{\link{incR.dataExample}}
was produced by an iButton® device (Maxim) and represents raw data to start the analysis
of incubation behaviour
}
\details{
see \code{\link{incR.prep}} for more details on the variables of this data set.
}
\keyword{datasets}

