% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR.prep.r
\name{incR.prep}
\alias{incR.prep}
\title{Data manipulation for incubation data analysis}
\usage{
incR.prep(data, date.name, date.format, timezone, temperature.name)
}
\arguments{
\item{data:}{raw data from incubation time series. It must contain a column 
with date + time information for each observation. The function is written to
handle date and time concatenated in one unique column (see example below).}

\item{date.name:}{name of the date + time column}

\item{date.format:}{date and time format for date + time column. 
It must be a character object as specified in the function \emph{strptime}.   
\emph{incR.prep} assumes the the date + time column contains date and time, 
If date and time are in different columns, please concatenate them in one
column before running the function (see example below).
its format accordingly.}

\item{timezone:}{time zone for time calculations. See \emph{strptime}
documentation for more details.}

\item{temperature.name:}{name of the column storing temperature information.}
}
\value{
The original data frame (data) with additional colunms for:
\enumerate{
\item index: a running number identifying every row in the data set.
\item dec.time: time in decimal hours. 
\item time: in  'H:M' format.
\item hour: in 'H' format.
\item year: in 'Y' format.
\item date: in  'Y-m-d' format.
\item temp1: difference between the ith temperature value (Ti) and the previous one (Ti-1).
}
}
\description{
Preparing incubation time series for further analysis. 
Initial data manipulation.
This function takes a data file containing a temporal series of temperature
and adds some the extra variables needed to use
further functions embedded in the incR package. 
It simply acommodates the raw data frame, reformating dates and times 
automatically.
}
\examples{
To be included
}
\author{
Pablo Capilla
}

