% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR.R
\docType{package}
\name{incR}
\alias{incR}
\alias{incR-package}
\title{incR: A package designed to analyse avian incubation
data.}
\description{
The functions contained in this package help the user to get
useful biological information from a raw time-series of temperatures.
}
\section{incR functions}{


\code{\link{incR_prep}}, \code{\link{incR_scan}},
\code{\link{incR_activity}}, \code{\link{incR_constancy}}
}

