% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR.scan.r
\name{incR.scan}
\alias{incR.scan}
\title{Calculation of incubation behaviour}
\usage{
incR.scan(data, lower.time = 22, upper.time = 3, maxinc.Temp,
  sensitivity = 0.15, time.dif = 20, maxNightVar_accepted,
  env.data = FALSE)
}
\arguments{
\item{data:}{data frame for analysis. It must contained four columns named as follow:
'date', 'temp1', 'dec.time' and 'index'. \code{\link{incR.prep}} returns a data frame
ready to be passed through \emph{incR.scan}}

\item{lower.time:}{lower limit of time window for calibration (numeric).}

\item{upper.time:}{upper limit of time window for calibration (numeric).}

\item{maxinc.Temp:}{maximum temperature of incubation.}

\item{sensitivity:}{ratio of reduction in temperature threshold. When nest temperature
does not drop close to environmental temperatures, this value can be kept to 1. If 
nest temperature follows environmental temperature, then adjustment of this value may
be required to detect short on/off-bouts.}

\item{time.dif:}{temperature difference between \emph{maxinc.Temp} and an observation which
triggers the sensitivity parameter.}

\item{maxNightVar_accepted:}{maximum temperature variation between two consecutive points
within the calibrating window that is accepted. If this variation value is surpassed, 
a previous night is used for calibration.}

\item{env.data:}{not yet supported}
}
\value{
The function returns a list with two objects. The first object is the original
data frame with an extra column named 'inc.vector'. This vector is formed by 1s and 0s,
representing whether the incubating individual is (1) or outside the nest (0).

The second object is a data frame with one day per row. Four columns tell the user
the thersholds employed to estimate incubating individual behaviuor. A fifth column accounts
for the ratio between the calibrating window temperature variation and the variation in temperature 
between 11am and 3pm. The lower this value the more clear the pattern between night and day
variation. It may serve the user for indication of the performance of the algorithm.
}
\description{
This is the core algorithm of \emph{incR} and classifies time points as 1s or 0s depeding on
whether or not the incubating individual is estimated to be on the eggs. 
The algorithm
uses night variation to daily-calibrate itself to temperature variation when the incubating
individual is (assumed to be) on the eggs. 
Therefore, a major assumption of this algorithm is that
there is a period of time in which temperature can be assumed to be constant or
representative of time windows of effective incubation. This time window is defined by
two arguments: lower.time and upper.time. The function is optimised to work using
a data frame produced by incR.prep.

In the near future, extended functionality will be included, especially including
environemtal temperature information into analysis. The
performance of this function has been evaluated in several bird species and geographic areas, 
but calibration using pilot data is recommended.
}
\details{
Description of the algorithm as it currently works. Future updates will
deal with environmental data.
}
\examples{
To be included
}
\author{
Pablo Capilla
}
\seealso{
\code{\link{incR.prep}} \code{\link{incR.constancy}} \code{\link{incR.activity}}
}

