% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inR.t.R
\name{incR.t}
\alias{incR.t}
\title{Calculation of temperature average and variance for customised
time windows}
\usage{
incR.t(data, limits = NULL, coor = NULL, activity.times = FALSE,
  civil.twilight = FALSE, time.zone = NULL)
}
\arguments{
\item{data}{data frame containing a time-series vector of 1s and 0s, where "1"
means "incubating individual inside nest" and "0" means "incubating individual 
outside the nests". This vector, 
under the name of "inc.vector", is provided by \code{\link{incR.scan}} in the 
first object of the returned list. A column named "date" is needed to refer to daily
calculations.}

\item{limits}{vector of length = 2 giving the time limits for calculations. For example,
'c(6,20)' would calculate temperature averages and variance for two time periods, from 6 to 20
and from 20 to 6 of the next day. 'civil.twilight' and 'activiy.times' must be
FALSE to allow the use of 'limits'.}

\item{coor}{coordinates for the location where temperature was recorded,
formatted as decimal degrees N/S, decimal degress E/W.
When 'civil.twilight' is TRUE, 'coor' allows the user to define sunrise and sunset times
based on the \code{\link{crepuscule}} function (in 'maptools' package).}

\item{activity.times}{TRUE or FALSE. Set as TRUE when time periods for calculation
are defined by \code{\link{incR.activity}}. Data must contain a column named 
'inc.vector' for the use of \code{\link{incR.activity}}.}

\item{civil.twilight}{TRUE or FALSE. Set as TRUE when time periods for calculation
are to be defined by civil twilight times - calculated using \emph{crepuscule{maptools}}. 
If civil.twilight = TRUE', 'coor' and 'time.zone' need to be specified.}

\item{time.zone}{time zone for \emph{crepuscule{maptools}} dawn and dusk calculations.}
}
\value{
a data frame containing onset and end of activity times for each day in \emph{data}.
}
\description{
Calculation of temperature average and variation between two customised
time periods per day. Originally, this function was thought to calculate nest temperature
between day and night. 'Day' and 'night' being defined either by the user, 
by activity times or by civil twilight times.
}
\examples{
# loading example data
data(incR.incubationExample)

# calculation based on chosen times from 6am to 7pm and 7pm to 6am
incR.t (data=incR.incubationExample, 
        limits=c(6,19), 
        coor=NULL, 
        civil.twilight=FALSE, 
        activity.times=FALSE,
        time.zone=NULL)
        
# calculation based on activity times
incR.t (data=incR.incubationExample, 
        limits=NULL, 
        coor=NULL, 
        civil.twilight=FALSE, 
        activity.times=TRUE,
        time.zone=NULL)
        
# calculation based on civil twilight
incR.t (data=incR.incubationExample, 
        limits=NULL, 
        coor=c(42,0.89), 
        civil.twilight=TRUE, 
        activity.times=FALSE,
        time.zone="GMT")
}
\author{
Pablo Capilla
}
\seealso{
\code{\link{incR.prep}} \code{\link{incR.scan}} \code{\link{incR.activity}}
\code{\link{crepuscule}}
}

