% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inR_activity.r
\name{incR_activity}
\alias{incR_activity}
\title{Calculation of onset and end of daily activity.}
\usage{
incR_activity(data, vector.presence = NULL)
}
\arguments{
\item{data:}{data frame containing a time-series vector of 1 and 0, where "1"
means "females inside nests" and "0" means "female" off the nests. This vector, 
under the name of "inc.vector" is provided by \code{\link{incR_scan}} in the 
first object of the returned list. A column named "date" is needed to refer to daily
calculations.}

\item{vector.presence:}{(character class) name of the vector containing information
for incubating individual presence/ausence of nests.}
}
\value{
data frame containing a column (vector) for incubating individual presence/
ausence of nests.
}
\description{
This functins uses the vector of female presence in nest
to calculate onset of daily activity (first off-bout of a day),
and end of daily activity (last on-bout of a day). A column for dates, named "date"
is needed in the data argument.
}
\examples{

}
\seealso{
\code{\link{incR_prep}} \code{\link{incR_scan}}
}

