% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR_constancy.r
\name{incR_constancy}
\alias{incR_constancy}
\title{Calculation of percentage of daily time spent in incubation.}
\usage{
incR_constancy(data, sampling.rate, vector.presence = "female")
}
\arguments{
\item{data:}{NA}

\item{sampling.rate:}{NA}

\item{vector.presence:}{NA}
}
\value{

}
\description{
not available
}
\examples{

}
\seealso{
\code{\link{incR_prep}} \code{\link{incR_scan}} \code{\link{incR_activity}}
}

