% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR_prep.r
\name{incR_prep}
\alias{incR_prep}
\title{Data manupilation for incubation data analysis}
\usage{
incR_prep(data, date.format, timezone)
}
\arguments{
\item{data:}{raw data from incubation time series. This must include two columns with the
following names:
\itemize{
 \item DATE - date and time collapsed in one columns, eg. "dd/mm/yyyy hh:mm". 
 Its format must be specified in \emph{format} (see below).
 \item valueT - temperature values.
}}

\item{format:}{format for DATE as defined in the function \emph{strptime}.
\emph{incR_prep} assumes the DATE contains date and time, so choose
its format accordingly.}

\item{timezone:}{time zone for time calculations. See \emph{strptime}
documentation for more details.}
}
\value{
The original data frame with additional colunms for:
\enumerate{
\item index: a running number identifying every row in the data set.
\item time: time in DATE and "hour"
\item date: date in DATE and "year"
\item dec.time: time in decimal format
\item temp1 and temp2: differential temperatures defines as
\deqn{T[i] - T[i-1]} and  \deqn{T[i] - T[i-2]}.
}
}
\description{
Preparing incubation time series for further analysis. 
Initial data manipulation.
This function takes a data file containing a temporal series of temperature
and adds all the extra variables needed to use
further functions embedded in the incR package. 
It simply acommodates the data frame reformating dates, times and 
so on.
}
\examples{
NULL
}

