% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR_rawdata.R
\docType{data}
\name{incR_rawdata}
\alias{incR_rawdata}
\title{An example of incubation temperature time-series}
\format{
A data frame with 954 rows and 2 variables, representing three days of nest
temperatures.
\describe{
  \item{DATE}{a date-time column. Both elements, date and time, are concatenated in one column}
  \item{temperature}{temperature recordings in Celsius}
}
}
\usage{
incR_rawdata
}
\description{
A dataset containing date, time and temperatures for a blue tit nest. This data set
was produced by an iButton device (Maxim Integrated) and 
represents raw data to start the analysis
of incubation behaviour.
}
\keyword{datasets}
