% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR_scan.r
\name{incR_scan}
\alias{incR_scan}
\title{Calculation of incubation behaviour.}
\usage{
incR_scan(data = data, lower.time = 22, upper.time = 3,
  env.data = FALSE, sensitivity = 0.15, time.dif = 20, MAX.T = 38,
  maxNightVar_accepted = 2)
}
\arguments{
\item{data:}{ny}

\item{lower.time:}{ny}

\item{upper.time:}{ny}

\item{env.data:}{ny}

\item{sensitivity:}{ny}

\item{time.dif:}{ny}

\item{MAX.T:}{ny}

\item{maxNightVar_accepted:}{ny}
}
\value{

}
\description{
This is the core algorithm of incR and classifies time points as 1s or 0s depeding on
whether or not the incubating individual is estimated to be on the eggs. 
The algorithm
uses night variation to calibrate itself to temperature variation when the incubating
individual is (assumed to be) on the eggs. 
Therefore, a major assumption of this algorithm is that
there is a period of time in which temperature can be assumed to be constant or
representative of time windows of effective incubation. This time window is defined by
two arguments: lower.time and upper.time. The function is optimised to work using
a data frame produced by incR_prep.

The algorithm can be adapted to different environmental circunstances. Its
performance has been evaluated in several species and geographic areas, but calibration
using pilot data is recommended.
}
\examples{

}
\seealso{
\code{\link{incR_prep}}
}

