% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incRdataExample2.R
\docType{data}
\name{incRdataExample2}
\alias{incRdataExample2}
\title{A second example of incubation temperature time-series after 
the use of \code{link{incRprep}}.}
\format{A data frame with 709 rows and 3 variables, representing three days of nest
temperatures.
\describe{
  \item{DATE}{a date + time column. Both elements are concatenated in one column}
  \item{valueT}{temperature recordings}
  \item{env.temp}{environmental temperature}
}}
\usage{
incRdataExample2
}
\description{
A dataset containing temperatures, time and date for a blue tit nest. This data set
was produced by an iButton® device (Maxim) and represents raw data to start the analysis
of incubation behaviour. Environmental temperature have been added as average per hour.
}
\keyword{datasets}

