% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incRenv.R
\name{incRenv}
\alias{incRenv}
\title{incRenv}
\usage{
incRenv(data.nest, data.env, env.temperature.name, env.date.name,
  env.date.format, env.timezone)
}
\arguments{
\item{data.nest}{data frame containing nest temperature recordings. It must have two compulsory
columns 'date' and 'hour' displaying dates and the hour of each observations. These two column
are provided if the user uses \code{\link{incRprep}} before.}

\item{data.env}{data frame containing environmental temperatures to be merge with nest temperature
records. Please, provide date and time of each observation in one unique column as requested 
for \code{\link{incRprep}}).}

\item{env.temperature.name}{name of the column containing temperature recording in the 
\emph{data.env} data frame.}

\item{env.date.name}{name of the column containing date and time in the 
\emph{data.env} data frame.}

\item{env.date.format}{format of \emph{env.date.name}. Similar to \code{\link{incRprep}}.}

\item{env.timezone}{time zone of the environmental recordings. Similar to \code{\link{incRprep}}.}
}
\value{
The original \emph{data.nest} with an additional colunms for hour-averaged environmental
temperature. This new variable is especially thought to serve as \emph{env.temp} in
\code{\link{incRscan}}.
}
\description{
Matching environmental and nest temperatures per hour of observation.
This functions takes a data frame with recordings of environmental temperature and another one
with nest temperature recordings and merges both to provide environmental temperatures 
per hour of nest temperature recordings. The user can do this work manually, however, incRenv
automates data preparation (in combination with \code{\link{incRprep}}) for the use of 
\code{\link{incRscan}}.
}
\details{
This functions is thought to be used after \code{\link{incRprep}}. 
The data frame returned by \code{\link{incRprep}} has all the columns for the used of
\code{\link{incRenv}}.
}
\examples{
data(incRenvironmentalData)  # environmental data
head (incRenvironmentalData)

data(incRincubationExample)  # nest data
head (incRincubationExample)

# the first step in to format the raw data using incRprep
new.data <- incRprep (data=incRdataExample,
                      date.name= "DATE",
                      date.format= "\%d/\%m/\%Y \%H:\%M",
                      timezone="GMT",
                      temperature.name="valueT")
                      
# then use incRenv to merge environmental data
new.data2 <- incRenv (data.nest = new.data,
                      data.env = incRenvironmentalData, 
                      env.temperature.name = "valueT", 
                      env.date.name = "DATE", 
                      env.date.format = "\%d/\%m/\%Y \%H:\%M", 
                      env.timezone = "GMT")
                     
head (new.data2, 3)
}
\author{
Pablo Capilla
}
\seealso{
\code{\link{incRprep}} \code{\link{incRscan}}
}

