% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR.constancy.r
\name{incR.constancy}
\alias{incR.constancy}
\title{Calculation of the percentage of daily time spent in incubation.}
\usage{
incR.constancy(data, sampling.rate, vector.presence)
}
\arguments{
\item{data:}{data frame containing a time-series vector of 1 and 0, where "1"
means "females inside nests" and "0" means "female" off the nests. This vector, 
under the name of "inc.vector" is provided by \code{\link{incR.scan}} in the 
first object of the returned list. A column named "date" is needed to refer to daily
calculations.}

\item{sampling.rate:}{time between two consecutive temperature recordings.}

\item{vector.presence:}{name of the column (vector) storing the
information about the presence/ausence of the incubating indiviual in the nest.}
}
\value{
Daily percentage (in 0 to 1 scale) of time in nest. Return in a 
data frame with one day per raw.
}
\description{
This function calculates % of day time spent incubating
based on the "inc.vector" produced by \code{\link{incR.prep}}.
Current versions do not discriminate day and night times.
}
\examples{
coming soon
}
\seealso{
\code{\link{incR.prep}} \code{\link{incR.scan}} \code{\link{incR.activity}}
}

