% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR.R
\docType{package}
\name{incR}
\alias{incR}
\alias{incR-package}
\title{incR: Analysis of animal incubation}
\description{
This packages is formed by a suite of R functions that help the user to get
useful biological information from raw time series data of 
incubation temperatures. It is specially thought to be of interest for
the study of uni-parental or intermittent incubating species.
So far, this package is written in R language to allow the potential
contribution from a wide R audience.
}
\section{Release notes and caveats}{

The current version of \code{\link{incR}} is in an early stage of development.
Documentation and code are actively being updated and bugs are likely to be
found. If so, please, send me your recommendations.

In the near future:
\enumerate{
\item \code{\link{incR.scan}} will be further improved
to include environmental temperatures into the analysis of incubation behaviour.
\item \code{\link{incR.bouts}} will be extended to produce off-bout-specific
data, giving temperature drop, duration and time per individual off-bout.
}
}

\section{incR functions}{

So far this package contains the following functions:
\code{\link{incR.prep}}, \code{\link{incR.scan}},
\code{\link{incR.activity}}, \code{\link{incR.constancy}}
\code{\link{incR.bouts}}, \code{\link{incR.t}}
}

