% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incR.R
\docType{package}
\name{incR}
\alias{incR}
\alias{incR-package}
\title{incR: Analysis of animal incubation}
\description{
This packages is formed by a suite of R functions that help the user to get
useful biological information from raw time-series data of 
incubation temperatures. It is thought to be of interest for
the study of uni-parental or intermittent incubating species.
}
\section{Suggested workflow}{

Check the package vignettes to find a suggested workflow
and calibrate the main function in \code{incR}.
}

\section{incR functions}{

The current version of incR contains the following functions and three example data sets:
\code{\link{incRprep}}, \code{\link{incRenv}},
\code{\link{incRscan}},
\code{\link{incRactivity}}, 
\code{\link{incRconstancy}}
\code{\link{incRbouts}} and 
\code{\link{incRt}}, along with three example data sets.
}

