% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inR.t.r
\name{incR.t}
\alias{incR.t}
\title{Calculation of temperature average and variation for customised
time windows.}
\usage{
incR.t(data, limits = NULL, coor, activity.times = FALSE,
  civil.twilight = FALSE, time.zone = NULL)
}
\arguments{
\item{data:}{data frame containing a time-series vector of 1s and 0s, where "1"
means "incubating individual inside nest" and "0" means "incubating individual 
outside the nests". This vector, 
under the name of "inc.vector", is provided by \code{\link{incR.scan}} in the 
first object of the returned list. A column named "date" is needed to refer to daily
calculations.}

\item{limits:}{vector of length = 2 giving the time limits for calculaitons. For example,
'c(6,20)" would calculate temperature averages and variation for two time periods, from 6 to 20
and from 20 to 6 of the next day. 'civil.twilight' and 'activiy.times' must be
FALSE to allow the use of 'limits'.}

\item{coor:}{coordenates for the location where temperature was recorded. 
When 'civil.twilight' is TRUE, 'coor' allows to define sunrise and sunset times
based on the \code{\link{crepuscule}} function (in maptools package).}

\item{civil.twilight:}{TRUE or FALSE. Set as TRUE when time periods for calculation
are to be defined by civil twilight times - calculated using \code{\link{crepuscule}}.If
'civil.twilight = TRUE', 'coor' and 'time.zone' need to be specified.}

\item{activity.times:}{TRUE or FALSE. Set as TRUE when time periods for calculation
are defined by \code{\link{incR.activity}}.}

\item{time.zone:}{time zone for \code{\link{crepuscule}} dawn and dusk calculations.}
}
\value{
a data frame containing onset and end of activity times for each day in \emph{data}.
}
\description{
Calculation of temperature average and variation between two customised
time periods per day. Originally, this function was thought to calculate temperature
of incubation between day and night. 'Day' and 'night' being defined either by the user, 
by activity times or by civil twilight times.
}
\examples{
To be included
}
\author{
Pablo Capilla
}
\seealso{
\code{\link{incR.prep}} \code{\link{incR.scan}} \code{\link{incR.activity}}
\code{\link{crepuscule}}
}

