% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incRactivity.R
\name{incRactivity}
\alias{incRactivity}
\title{Calculation of onset and end of daily activity}
\usage{
incRactivity(data, vector.incubation)
}
\arguments{
\item{data}{data frame containing a time-series vector of 1s and 0s, where "1"
means "incubating individual inside nest" and "0" means "incubating individual 
outside the nests". This vector, 
under the name of "inc.vector" is provided by \code{\link{incRscan}} in the 
first object of the returned list. A column named "date" is needed to refer to daily
calculations.}

\item{vector.incubation}{(character class) name of the vector containing information
for incubating individual presence/absence of nests. \code{\link{incRscan}} produces
this vector named "inc.vector".}
}
\value{
a data frame containing onset and end of activity times for each day in \emph{data}.
}
\description{
Using a vector of incubating individual presence in nest, 
\emph{incRactivity}
calculates onset of activity (first off-bout in the morning),
and end of daily activity (last on-bout in the evening) per day. A column for dates, named "date"
is needed in the data argument.
}
\examples{
#' # loading example data
data(incRincubationExample)
incRactivity (data=incRincubationExample, 
               vector.incubation="inc.vector")
}
\author{
Pablo Capilla
}
\seealso{
\code{\link{incRprep}} \code{\link{incRscan}}
}

